fix: extracted apply to shared method and fixed context dependency
